+++ kikuchipy/projections/gnomonic_projection.py
@@ -16,17 +16,28 @@
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
 from typing import Union
+import warnings
 
 import numpy as np
 from orix.vector import Vector3d
 
 from kikuchipy.projections.spherical_projection import SphericalProjection
+from kikuchipy._util import deprecated
 
 
 class GnomonicProjection(SphericalProjection):
-    """Gnomonic projection of a vector as implemented in MTEX."""
+    """[*Deprecated*] Gnomonic projection of a vector as implemented in
+    MTEX.
+
+    .. deprecated:: 0.8.0
+
+        This class is deprecated and will be removed in 0.9.0, since it
+        is not used internally. If you depend on this class, please open
+        an issue at https://github.com/pyxem/kikuchipy/issues.
+    """
 
     @classmethod
+    @deprecated(since="0.8.0", removal="0.9.0")
     def vector2xy(cls, v: Union[Vector3d, np.ndarray]) -> np.ndarray:
         r"""Convert from 3D cartesian coordinates :math:`(x, y, z)` to
         2D Gnomonic coordinates :math:`(x_g, y_g)`.
@@ -50,7 +61,9 @@ class GnomonicProjection(SphericalProjection):
         >>> v = np.random.random_sample(30).reshape((10, 3))
         >>> xy = GnomonicProjection.vector2xy(v)
         """
-        polar_coordinates = super().vector2xy(v)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            polar_coordinates = super().vector2xy(v)
         polar, azimuth = polar_coordinates[..., 0], polar_coordinates[..., 1]
 
         # Map to upper hemisphere
@@ -71,6 +84,7 @@ class GnomonicProjection(SphericalProjection):
         return gnomonic
 
     @staticmethod
+    @deprecated(since="0.8.0", removal="0.9.0")
     def xy2vector(xy: np.ndarray) -> Vector3d:
         r"""Convert from 2D Gnomonic coordinates :math:`(x_g, y_g)` to
         3D cartesian coordiantes :math:`(x, y, z)`.